refactor(mina.ts): move transaction validation functions to separate file
